refactor(cfd-2d): Decompose ns_fvm/solver/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
